requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
requests
gradio
